requirements.txt
langchain
langchain_openai
langchain_huggingface
langchain_community
faiss-cpu
colorama
openai
huggingface
pypdf
gradio
